requirements.txt
fake-bpy-module
gTTS
openai
openai-whisper
pydub
reachy-sdk
requests
scipy
sounddevice